% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ewb_colors}
\alias{ewb_colors}
\title{Data frame containing Epi-Workbench brand standard colors}
\format{
A data frame with 5 rows and 4 columns:
\describe{
\item{group}{Brand/project group}
\item{subgroup}{Color palette subgroup}
\item{hex}{Hexidecimal code for the color}
\item{description}{Color description}
}
}
\usage{
ewb_colors
}
\description{
This is the code to create a data frame containing Epi-Workbench brand standard colors.
}
\keyword{datasets}
