% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_hidden_code_block.R
\name{insert_hidden_code_block}
\alias{insert_hidden_code_block}
\alias{hidden_code_block_template}
\title{Insert an Hidden Code Block into a Quarto Lesson}
\usage{
insert_hidden_code_block()

hidden_code_block_template()
}
\value{
Invisibly returns the code block text as a single character string.
}
\description{
This add-in inserts the standard "hidden" code block template
used for Epi-Workbench lessons at the current cursor location in RStudio.
}
\examples{
if (interactive()) insert_hidden_code_block()
}
\references{
For more information on hidden code blocks see: https://github.com/epi-workbench/EWBTemplates/wiki/Code-Blocks
}
\seealso{
Other Insert Addins: 
\code{\link{insert_hints}()},
\code{\link{insert_initial_code_block}()},
\code{\link{insert_lesson_yaml_header}()},
\code{\link{insert_noeval_code_block}()},
\code{\link{insert_solution_code_block}()},
\code{\link{insert_test_code_block}()}
}
\concept{Insert Addins}
