% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_hints.R
\name{insert_hints}
\alias{insert_hints}
\alias{hints_template}
\title{Insert Code Block Hints into a Quarto Lesson}
\usage{
insert_hints()

hints_template()
}
\value{
Invisibly returns the code block hint text as a single character string.
}
\description{
This add-in inserts the standard code block hint template
used for Epi-Workbench lessons at the current cursor location in RStudio.
}
\examples{
if (interactive()) insert_hints()
}
\references{
For more information on hints see: https://github.com/epi-workbench/EWBTemplates/wiki/Hints
}
\seealso{
Other Insert Addins: 
\code{\link{insert_hidden_code_block}()},
\code{\link{insert_initial_code_block}()},
\code{\link{insert_lesson_yaml_header}()},
\code{\link{insert_noeval_code_block}()},
\code{\link{insert_solution_code_block}()},
\code{\link{insert_test_code_block}()}
}
\concept{Insert Addins}
