% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_initial_code_block.R
\name{insert_initial_code_block}
\alias{insert_initial_code_block}
\alias{initial_code_block_template}
\title{Insert an Initial Code Block into a Quarto Lesson}
\usage{
insert_initial_code_block()

initial_code_block_template()
}
\value{
Invisibly returns the code block text as a single character string.
}
\description{
This add-in inserts the standard "initial" code block template used for
Epi-Workbench lessons at the current cursor location in RStudio.
}
\examples{
if (interactive()) insert_initial_code_block()
}
