% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_lesson_yaml_header.R
\name{insert_lesson_yaml_header}
\alias{insert_lesson_yaml_header}
\alias{lesson_yaml_header}
\title{Insert a YAML Header into a Quarto Lesson}
\usage{
insert_lesson_yaml_header()

lesson_yaml_header()
}
\value{
Invisibly returns the YAML header text as a single character
string.
}
\description{
This add-in inserts the standard YAML header used for
Epi-Workbench lessons at the current cursor location in RStudio.
}
\examples{
if (interactive()) insert_lesson_yaml_header()
}
