% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_noeval_code_block.R
\name{insert_noeval_code_block}
\alias{insert_noeval_code_block}
\alias{noeval_code_block_template}
\title{Insert an No Evaluation Code Block into a Quarto Lesson}
\usage{
insert_noeval_code_block()

noeval_code_block_template()
}
\value{
Invisibly returns the code block text as a single character string.
}
\description{
This add-in inserts the standard "noeval" code block template
used for Epi-Workbench lessons at the current cursor location in RStudio.
}
\examples{
if (interactive()) insert_noeval_code_block()
}
