% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_solution_code_block.R
\name{insert_solution_code_block}
\alias{insert_solution_code_block}
\alias{solution_code_block_template}
\title{Insert an Solution Code Block into a Quarto Lesson}
\usage{
insert_solution_code_block()

solution_code_block_template()
}
\value{
Invisibly returns the code block text as a single character string.
}
\description{
This add-in inserts the standard "solution" code block template
used for Epi-Workbench lessons at the current cursor location in RStudio.
}
\examples{
if (interactive()) insert_solution_code_block()
}
