% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_test_code_block.R
\name{insert_test_code_block}
\alias{insert_test_code_block}
\alias{test_code_block_template}
\title{Insert an Test Code Block into a Quarto Lesson}
\usage{
insert_test_code_block()

test_code_block_template()
}
\value{
Invisibly returns the code block text as a single character string.
}
\description{
This add-in inserts the standard "test" code block template
used for Epi-Workbench lessons at the current cursor location in RStudio.
}
\examples{
if (interactive()) insert_test_code_block()
}
