% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_colors}
\alias{my_colors}
\title{Data frame of colors to reuse}
\format{
A data frame with 20 rows and 4 columns:
\describe{
\item{group}{Brand/project group}
\item{subgroup}{Brand/project subgroup}
\item{hex}{Hexidecimal code for the color}
\item{description}{Color description}
}
}
\usage{
my_colors
}
\description{
This is the code to create a data frame with color themes that I want to
}
\keyword{datasets}
