% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_hints.R
\name{insert_hints}
\alias{insert_hints}
\alias{hints_template}
\title{Insert Code Block Hints into a Quarto Lesson}
\usage{
insert_hints()

hints_template()
}
\value{
Invisibly returns the code block hint text as a single character string.
}
\description{
This add-in inserts the standard code block hint template
used for Epi-Workbench lessons at the current cursor location in RStudio.
}
\examples{
if (interactive()) insert_hints()
}
